update separação de funções
