+++ pag/teste.md
@@ -1,0 +1,19 @@
+# Markdown Links
+
+## 1. Prefácio
+
+[Markdown](https://pt.wikipedia.n) é uma linguagem de marcação
+muito popular entre os programadores. É usada em muitas plataformas que
+manipulam texto (GitHub, fórum, blogs e etc) e é muito comum encontrar arquivos
+com este formato em qualquer repositório (começando pelo tradicional
+`README.md`).
+
+Os arquivos `Markdown` normalmente contém _links_ que podem estar
+quebrados, ou que já não são válidos, prejudicando muito o valor da
+informação que está ali.
+
+Uma comunidade open source nos propôs criar uma ferramenta, usando
+[Node.js](https://nodejs.org/), que leia e analise arquivos no formato
+`Markdown`, para verificar os arquivos que contenham links e mostrar algumas
+estatísticas.
+
